requirements.txt
fastapi[standard]
python-multipart
uvicorn[standard]
python-dotenv
motor
pymongo
bcrypt
python-jose[cryptography]
passlib[bcrypt]